rename pocket service inner files
