+++ src/services/pocket/index.ts
@@ -1,11 +1,11 @@
 import type { Service } from 'utils/services';
 import { name, getQueryKey } from './const';
-import { forceGet, get, search } from './apiGet';
-import { authorize } from './apiAuthorize';
-import { connect } from './apiConnect';
-import { disconnect } from './apiDisconnect';
+import { forceGet, get, search } from './get';
+import { authorize } from './authorize';
+import { connect } from './connect';
+import { disconnect } from './disconnect';
 import { isConnected } from './helpers';
-import { archiveItem, deleteItem } from './apiActions';
+import { archiveItem, deleteItem } from './actions';
 import { Icon } from './Icon';
 
 export type PocketServiceName = 'pocket';
